requirements.txt
asgiref==3.5.2
attrs==22.1.0
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
Django==4.0.5
django-storages==1.13.2
djangorestframework==3.14.0
drf-yasg==1.21.4
exceptiongroup==1.0.4
filelock==3.7.1
FirstApp==1.3.0
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
Pillow==9.4.0
platformdirs==2.5.2
pluggy==1.0.0
protobuf==4.21.12
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytest==7.2.0
pytz==2022.7
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.14.1
